corregido bug que eliminaba todos los textos cuando salias del debug
git-svn-id: https://svn.code.sf.net/p/warcom/code/trunk/Warcom_Game_Engine@477 bdfe1ff2-85b1-43ac-a073-3d879cfd3ae1

+++ debug.prg
@@ -114,7 +114,9 @@ begin
 		if (not debugMode && actDebugMode)
 			//limpiamos los textos
 			for (i=0;i<cMaxDebugInfo;i++)
-				delete_text(idDebugText[i]);
+				if (idDebugText[i] <> 0 )
+					delete_text(idDebugText[i]);
+				end;
 			end;
 			//Quitamos al player el blend aditivo para ver las colisiones
 			if (idPlayer<>0) idPlayer.flags &= ~ B_ABLEND; end;
